Time: 5 ms (19.97%), Space: 8.6 MB (53.83%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        int ans=0;
+        long long ans=0;
         while(x){
             int digi=x%10;
             if(ans<INT_MIN /10||ans>INT_MAX /10)
@@ -9,6 +9,7 @@ public:
             ans=ans*10+digi;
             x=x/10;
         }
+        
         
         return ans;
     }
